updated controllers,helpers and endpoints

+++ Backend/Database/Stored Procedures/Comments/DeleteCommentByUserId.sql
@@ -1,9 +1,14 @@
+USE StackOverflow
+GO
+
 CREATE OR ALTER PROCEDURE sp_DeleteCcommentbyUserId
-  @user_id VARCHAR(255)
+  @userId VARCHAR(255)
 AS
 BEGIN
-  DELETE FROM comments
-  WHERE user_id = @user_id 
+  DELETE FROM comments WHERE user_id = @userId
+
+  SELECT * FROM comments 
+ 
 END
 
-EXEC  sp_DeleteCcommentbyUserId @user_id='U002';
+-- EXEC  sp_DeleteCcommentbyUserId @user_id='U002';
